+++ custom_components/wattpilot/number.py
@@ -20,7 +20,7 @@ from .descriptions import (
     WattpilotNumberEntityDescription,
 )
 from .entities import ChargerPlatformEntity, filter_descriptions
-from .utils import async_SetChargerProp
+from .utils import async_SetChargerProp, register_push_entity
 
 if TYPE_CHECKING:
     from homeassistant.helpers.entity_platform import AddEntitiesCallback
@@ -57,7 +57,7 @@ async def async_setup_entry(
             continue
         entities.append(entity)
         if entity._source == SOURCE_PROPERTY:
-            push_entities[entity._identifier] = entity
+            register_push_entity(push_entities, entity._identifier, entity)
 
     _LOGGER.info(
         "%s - async_setup_entry: setup %s %s entities",