+++ custom_components/wattpilot/update.py
@@ -28,6 +28,7 @@ from .descriptions import (
 from .entities import ChargerPlatformEntity, filter_descriptions
 from .utils import (
     GetChargerProp,
+    register_push_entity,
 )
 
 if TYPE_CHECKING:
@@ -66,7 +67,7 @@ async def async_setup_entry(
             continue
         entities.append(entity)
         if entity._source == SOURCE_PROPERTY:
-            push_entities[entity._identifier] = entity
+            register_push_entity(push_entities, entity._identifier, entity)
 
     _LOGGER.info(
         "%s - async_setup_entry: setup %s %s entities",